// pg-frontend/src/generated/models/Question.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Question = {
    acceptedNum?: number;
    answer?: string;
    content?: string;
    createTime?: string;
    difficulty?: string;
    id?: number;
    isDelete?: number;
    judgeCase?: string;
    judgeConfig?: string;
    license?: string;
    packageType?: string;
    referenceLanguage?: string;
    referenceSolution?: string;
    source?: string;
    sourceUrl?: string;
    submitNum?: number;
    status?: string;
    tags?: string;
    title?: string;
    updateTime?: string;
    userId?: number;
};